# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/wearable-data": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List Wearable Data Records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WearableData"
                ],
                "summary": "List Wearable Data Records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Recorded Timestamp",
                        "name": "recorded_timestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.GetAllWearableDataRes"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Wearable Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WearableData"
                ],
                "summary": "Create Wearable Data",
                "parameters": [
                    {
                        "description": "body for WearableDataCreate",
                        "name": "wearabledata",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.WearableDataCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wearable data created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/wearable-data/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Wearable Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WearableData"
                ],
                "summary": "Get Wearable Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wearable Data ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.WearableDataRes"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Wearable Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WearableData"
                ],
                "summary": "Update Wearable Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wearable Data ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "body for WearableDataUpdate",
                        "name": "wearabledata",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.WearableDataUpdateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wearable data updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Wearable Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WearableData"
                ],
                "summary": "Delete Wearable Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wearable Data ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wearable data deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "health_sync.GetAllWearableDataRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "wearable_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/health_sync.WearableDataRes"
                    }
                }
            }
        },
...
        "health_sync.WearableDataCreate": {
            "type": "object",
            "properties": {
                "calories_burned": {
                    "type": "number"
                },
                "device_type": {
                    "type": "string"
                },
                "distance_traveled": {
                    "type": "number"
                },
                "heart_rate": {
                    "type": "integer"
                },
                "recorded_timestamp": {
                    "type": "string"
                },
                "sleep_duration": {
                    "type": "number"
                },
                "step_count": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "workout_type": {
                    "type": "string"
                }
            }
        },
        "health_sync.WearableDataRes": {
            "type": "object",
            "properties": {
                "calories_burned": {
                    "type": "number"
                },
                "device_type": {
                    "type": "string"
                },
                "distance_traveled": {
                    "type": "number"
                },
                "heart_rate": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "recorded_timestamp": {
                    "type": "string"
                },
                "sleep_duration": {
                    "type": "number"
                },
                "step_count": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                },
                "workout_type": {
                    "type": "string"
                }
            }
        },
        "health_sync.WearableDataUpdateBody": {
            "type": "object",
            "properties": {
                "calories_burned": {
                    "type": "number"
                },
                "device_type": {
                    "type": "string"
                },
                "distance_traveled": {
                    "type": "number"
                },
                "heart_rate": {
                    "type": "integer"
                },
                "recorded_timestamp": {
                    "type": "string"
                },
                "sleep_duration": {
                    "type": "number"
                },
                "step_count": {
                    "type": "integer"
                },
                "workout_type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}